+++ src/components/InfoTooltip/index.tsx
@@ -1,11 +1,17 @@
 import { FloaterPlacement, Icon, Tooltip } from "@stellar/design-system";
 import "./styles.scss";
 
+const ICONS = {
+  info: <Icon.Info />,
+  check: <Icon.CheckCircle color="green" />,
+};
+
 interface InfoTooltipProps {
   children: React.ReactNode;
   infoText: React.ReactNode;
   placement?: FloaterPlacement;
   hideTooltip?: boolean;
+  icon?: "info" | "check";
 }
 
 export const InfoTooltip = ({
@@ -13,20 +19,19 @@ export const InfoTooltip = ({
   infoText,
   placement = "right",
   hideTooltip = false,
+  icon = "info",
 }: InfoTooltipProps) => {
   if (hideTooltip) {
     return <>{children}</>;
   }
 
+  const iconComponent = ICONS[icon];
+
   return (
     <div className="InfoTooltip">
       {children}
       <Tooltip
-        triggerEl={
-          <div className="InfoTooltip__button">
-            <Icon.Info />
-          </div>
-        }
+        triggerEl={<div className="InfoTooltip__button">{iconComponent}</div>}
         placement={placement}
       >
         {infoText}